prep src file structure for styleguidist
