+++ src/components/imageEffects.js
